@@ -19,25 +19,24 @@
   ],
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.7",
-    "@angular/cdk": "^6.1.7",
-    "@angular/common": "^6.1.7",
-    "@angular/core": "^6.1.7",
-    "@angular/forms": "^6.1.7",
-    "@angular/http": "^6.1.7",
-    "@angular/material": "^6.1.7",
-    "@angular/platform-browser": "^6.1.7",
-    "@angular/platform-browser-dynamic": "^6.1.7",
-    "@angular/router": "^6.1.7",
-    "@auth0/angular-jwt": "^2.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^3.2.0",
-    "@ng-select/ng-select": "^2.20.5",
-    "@tinymce/tinymce-angular": "^3.0.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^5.3.0",
+    "@ng-select/ng-select": "^3.7.3",
+    "@tinymce/tinymce-angular": "^3.5.0",
     "@types/geojson": "^7946.0.2",
-    "@types/leaflet": "^1.2.6",
-    "bootstrap": "^4.1.3",
-    "codelyzer": "^4.4.4",
-    "core-js": "^2.5.3",
+    "@types/leaflet": "^1.3.4",
+    "bootstrap": "^4.4.1",
+    "classlist.js": "1.1.20150312",
+    "codelyzer": "^4.5.0",
+    "core-js": "^3.1.3",
     "hammerjs": "^2.0.8",
     "jszip": "^3.2.2",
     "keycloak-angular": "3.0.3",
@@ -46,40 +45,42 @@
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.14",
     "mygovbc-bootstrap-theme": "^0.4.1",
+    "ng-zorro-antd": "8.5.2",
     "ng2-bootstrap-modal": "^1.0.1",
-    "ngx-cookie-service": "^1.0.10",
-    "ng-zorro-antd": "1.8.1",
+    "ngx-cookie-service": "^3.0.3",
     "path": "^0.12.7",
-    "rxjs": "^6.3.2",
-    "rxjs-compat": "^6.3.2",
-    "tinymce": "^5.0.4",
-    "tslint": "^5.11.0",
-    "web-animations-js": "^2.3.1",
-    "xlsx": "^0.14.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.4",
+    "rxjs-compat": "^6.5.4",
+    "tinymce": "^5.2.0",
+    "tslint": "^5.19.0",
+    "web-animations-js": "^2.3.2",
+    "xlsx": "^0.15.6",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.8.1",
-    "@angular-devkit/schematics": "0.0.40",
-    "@angular/cli": "^6.2.1",
-    "@angular/compiler": "^6.1.7",
-    "@angular/compiler-cli": "^6.1.7",
+    "@angular-devkit/build-angular": "^0.803.23",
+    "@angular-devkit/core": "^8.3.25",
+    "@angular-devkit/schematics": "8.3.25",
+    "@angular/cli": "^8.3.25",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
     "@types/arcgis-js-api": "4.6.0",
     "@types/jasmine": "2.8.6",
     "@types/lodash": "4.14.106",
-    "@types/node": "~6.0.60",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
+    "@types/node": "~10.17.17",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
     "pre-push": "^0.1.1",
-    "protractor": "~5.3.0",
-    "puppeteer": "^2.0.0",
-    "ts-node": "2.0.0",
-    "typescript": "^2.9.2"
+    "protractor": "~5.4.3",
+    "puppeteer": "^2.1.1",
+    "ts-node": "7.0.1",
+    "typescript": "~3.4.5",
+    "tree-kill": "^1.2.2"
   }
 }
