@@ -1,8 +1,5 @@
-import { Component, OnInit, SimpleChanges, HostListener } from '@angular/core';
+import { Component, OnInit, HostListener } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { Dictionary } from 'lodash';
-import * as moment from 'moment';
-import { Utils } from 'app/shared/utils/utils';
 
 @Component({
   selector: 'app-input-modal',
@@ -14,8 +11,7 @@ export class InputModalComponent implements OnInit {
   groupName = '';
 
   constructor(
-    public activeModal: NgbActiveModal, // also used in template
-    private utils: Utils
+    public activeModal: NgbActiveModal // also used in template
   ) { }
 
   ngOnInit() {
@@ -34,7 +30,7 @@ export class InputModalComponent implements OnInit {
   }
 
   // Handle escape key press.
-  @HostListener('document:keydown.escape', ['$event']) onKeydownHandler(event: KeyboardEvent) {
+  @HostListener('document:keydown.escape', []) onKeydownHandler() {
     this.cancel();
   }
 }
