@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import { DocumentService } from 'app/services/document.service';
@@ -12,7 +12,7 @@ export class DocumentDetailResolver implements Resolve<Observable<object>> {
     private configService: ConfigService
   ) { }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<object> {
+  resolve(route: ActivatedRouteSnapshot): Observable<object> {
     const docId = route.paramMap.get('docId');
     // We need to make sure that getLists populates the configService. Then just returns the documentService getById
     return this.configService.getLists()
