@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import { SearchService } from 'app/services/search.service';
@@ -28,8 +28,7 @@ export class ProjectResolver implements Resolve<Object> {
       {_id: projId},
       true,
       {},
-      '',
-      true
+      ''
     );
   }
 }
