@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
 import { TopicService } from 'app/services/topic.service';
@@ -10,7 +10,7 @@ export class TopicResolver implements Resolve<Observable<object>> {
     private topicService: TopicService
   ) { }
 
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<object> {
+  resolve(route: ActivatedRouteSnapshot): Observable<object> {
     const pageNum = Number(route.queryParams['pageNum'] ? route.queryParams['pageNum'] : 1);
     const pageSize = Number(route.queryParams['pageSize'] ? route.queryParams['pageSize'] : 1000);
     return this.topicService.getAllTopics(pageNum, pageSize);
