@@ -1,4 +1,4 @@
-import { Component, Input, Output, OnInit, EventEmitter, OnDestroy } from '@angular/core';
+import { Component, Input, Output, OnInit, EventEmitter } from '@angular/core';
 
 import { TableComponent } from 'app/shared/components/table-template/table.component';
 import { TableObject } from 'app/shared/components/table-template/table-object';
@@ -50,7 +50,7 @@ export class LinkOrganizationTableRowsComponent implements OnInit, TableComponen
       if (item.checkbox) {
         this.storageService.state.selectedOrgs.push(item);
       } else {
-        this.storageService.state.selectedOrgs = this.storageService.state.selectedOrgs.filter(function (value, index, arr) {
+        this.storageService.state.selectedOrgs = this.storageService.state.selectedOrgs.filter(function (value) {
           return value._id !== item._id;
         });
       }
