@@ -19,35 +19,32 @@
  */
 
 /** IE9, IE10 and IE11 requires all of the following polyfills. **/
-import 'core-js/es6/symbol';
-import 'core-js/es6/object';
-import 'core-js/es6/function';
-import 'core-js/es6/parse-int';
-import 'core-js/es6/parse-float';
-import 'core-js/es6/number';
-import 'core-js/es6/math';
-import 'core-js/es6/string';
-import 'core-js/es6/date';
-import 'core-js/es6/array';
-import 'core-js/es6/regexp';
-import 'core-js/es6/map';
-import 'core-js/es6/set';
+import 'core-js/features/symbol';
+import 'core-js/features/function';
+import 'core-js/features/parse-int';
+import 'core-js/features/parse-float';
+import 'core-js/features/number';
+import 'core-js/features/math';
+import 'core-js/features/string';
+import 'core-js/features/date';
+import 'core-js/features/array';
+import 'core-js/features/regexp';
+import 'core-js/features/map';
+import 'core-js/features/set';
 
-import 'core-js/es6/weak-map';
-import 'core-js/es6/weak-set';
-import 'core-js/es7/object';
-import 'core-js/es7/array';
+import 'core-js/features/weak-map';
+import 'core-js/features/weak-set';
+import 'core-js/features/object';
 
 /** IE10 and IE11 requires the following for NgClass support on SVG elements */
-// import 'classlist.js';  // Run `npm install --save classlist.js`.
+import 'classlist.js';  // Run `npm install --save classlist.js`.
 
 /** IE10 and IE11 requires the following to support `@angular/animation`. */
 import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 
 
 /** Evergreen browsers require these. **/
-import 'core-js/es6/reflect';
-import 'core-js/es7/reflect';
+import 'core-js/features/reflect';
 
 
 /** ALL Firefox browsers require the following to support `@angular/animation`. **/
@@ -66,12 +63,8 @@ import 'zone.js/dist/zone';  // Included with Angular CLI.
  * APPLICATION IMPORTS
  */
 
-/**
- * Date, currency, decimal and percent pipes.
- * Needed for: All but Chrome, Firefox, Edge, IE11 and Safari 10
- */
-// import 'intl';  // Run `npm install --save intl`.
-
+ // global is not defined fix (https://github.com/Leonidas-from-XIV/node-xml2js/issues/538)
+(window as any)['global'] = window;
 
 // Specific polyfill needed for IE11 using current packages as of Sep 24, 2018:
 // https://github.com/uxitten/polyfill/blob/master/string.polyfill.js
